debut de la liaision du backend et du frontend

+++ backend/src/main/java/com/docuai/ai/dto/AiModelConfigDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class AiModelConfigDto {
@@ -17,7 +18,7 @@ public class AiModelConfigDto {
         private UUID id;
         private String provider;
         private String modelName;
-        private Double temperature;
+        private BigDecimal temperature;
         private Integer maxTokens;
         private Boolean isDefault;
     }
@@ -30,7 +31,7 @@ public class AiModelConfigDto {
         private String provider;
         private String modelName;
         private String apiKey; // Plaintext passing in, will be encrypted in DB
-        private Double temperature;
+        private BigDecimal temperature;
         private Integer maxTokens;
         private Boolean isDefault;
     }
